This commit is partialy functional. I think there are some files that are probably missing. This is a very significant update so expect things to look very different.
Known issues:
 - GUI freezes when searching for a network.
 - Timezone set interface is very basic and probably not functional for now.
  - Set volume doesn't work yet
  - App store doesn't work yet other than the No WiFi connection page.
And probably more! This was just a first significant update with LVGL. I do still need to make an installer but since the file structure is constantly changing, and there are potential other files that may need to be added, I may wait until the project is more stable.